skills/fabius-cohors/references/agents/python/retail-ai-location-strategy/pyproject.toml
[project]
name = "retail-ai-location-strategy"
version = "1.0.0"
description = "Multi-agent AI pipeline for retail site selection using Google ADK"
authors = [
    {name = "Lavi Nigam", email = "lavinigam@google.com"},
    {name = "Deepak Moonat", email = "dmoonat@google.com"},
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10,<3.13"
keywords = ["adk", "agents", "gemini", "retail", "location-strategy"]

dependencies = [
    "google-adk>=1.20.0",
    "googlemaps>=4.10.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.36.0",
]

[project.urls]
Repository = "https://github.com/google/adk-samples"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.32.0",
    "ruff>=0.4.6",
    "mypy>=1.18.2",
    "codespell>=2.4.1",
]

[tool.ruff]
# Point back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint.isort]
# Categorizes the local package for correct import sorting
known-first-party = ["app", "retail_ai_location_strategy"]

[tool.ruff.lint.per-file-ignores]
# Ignore imports not at top of file, unused imports, and import sorting in __init__.py
"__init__.py" = ["E402", "F401", "I001"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv", "app/frontend"]

[tool.hatch.build.targets.wheel]
# Maps the project to the 'app' directory
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.agent-starter-pack]
base_template = "adk_base"
description = "A multi-agent AI pipeline for retail site selection that uses Gemini to research markets, map competitors, analyze gaps, and generate strategic recommendations with visual reports"
example_question = "I want to open a coffee shop in Indiranagar, Bangalore"

[tool.agent-starter-pack.settings]
deployment_targets = ["agent_engine", "cloud_run"]
interactive_command = "dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"